refactor(relaxtime): split transport coefficient validation layer
