+++ docs/dev/active/2026-03-27_src-P1P2补遗任务单.md
@@ -80,13 +80,13 @@
 - 体检 3.5：`TransportCoefficients`/`AverageScatteringRate`/`TransportWorkflow` 体量大。
 
 **任务**
-- [ ] 仅做第一步拆分方案（不大改逻辑）：确定拆分边界与最小迁移顺序。
-- [ ] 优先选 1 个模块做“低风险拆分试点”（建议 `TransportCoefficients` 的校验/适配层）。
-- [ ] 保持公共 API 不变并补最小回归。
+- [x] 仅做第一步拆分方案（不大改逻辑）：确定拆分边界与最小迁移顺序（本轮边界：将 `TransportCoefficients` 校验职责拆至独立模块）。
+- [x] 优先选 1 个模块做“低风险拆分试点”（建议 `TransportCoefficients` 的校验/适配层）。
+- [x] 保持公共 API 不变并补最小回归。
 
 **验收**
-- [ ] 文件职责更清晰。
-- [ ] 行为与测试结果不回归。
+- [x] 文件职责更清晰。
+- [x] 行为与测试结果不回归。
 
 ## 3. 优先级顺序（建议）
