+++ src/relaxtime/TransportCoefficients.jl
@@ -34,8 +34,13 @@ using Main.PNJLQuarkDistributions_Aniso: quark_distribution_aniso, antiquark_dis
 
 using Main.ParameterTypes: QuarkParams, ThermoParams
 using Main.ParameterAdapters: normalize_quark_input, normalize_thermo_input
-using Main.ValidationUtils: require_finite, require_positive_finite,
-    require_nonnegative_finite, require_positive_integer, validate_grid_weight_pair
+using Main.ValidationUtils: require_positive_finite, require_positive_integer, validate_grid_weight_pair
+using ..TransportCoefficientsValidation: _validate_tau_namedtuple,
+    _validate_quark_thermo_inputs,
+    _validate_bulk_coeffs_isentropic,
+    _validate_transport_inputs,
+    _validate_conserved_charge_symbol,
+    _validate_diffusion_background
 
 export shear_viscosity, bulk_viscosity, bulk_viscosity_isentropic, electric_conductivity, transport_coefficients
 export diffusion_coefficient, diffusion_matrix
@@ -749,83 +754,6 @@ end
     return _merge_transport_integration_config(base_config, kwargs)
 end
 
-@inline function _validate_tau_namedtuple(tau::NamedTuple)
-    for sp in _SPECIES_ALL
-        τ = tau_for_species(sp, tau)
-        require_nonnegative_finite("tau.:$sp", τ)
-    end
-    return nothing
-end
-
-@inline function _validate_quark_thermo_inputs(
-    quark_params::NamedTuple,
-    thermo_params::NamedTuple;
-    provider=nothing,
-)
-    require_positive_finite("thermo_params.T", thermo_params.T)
-    require_finite("thermo_params.Φ", thermo_params.Φ)
-    require_finite("thermo_params.Φbar", thermo_params.Φbar)
-
-    check_mass = provider === nothing || !hasproperty(provider, :mass_for_species)
-    check_mu = provider === nothing || !hasproperty(provider, :mu_for_species)
-    for fl in _FLAVORS
-        m = getproperty(quark_params.m, fl)
-        μ = getproperty(quark_params.μ, fl)
-        if check_mass
-            require_nonnegative_finite("quark_params.m.$fl", m)
-        end
-        if check_mu
-            require_finite("quark_params.μ.$fl", μ)
-        end
-    end
-    return nothing
-end
-
-@inline function _validate_bulk_coeffs_isentropic(coeffs::NamedTuple)
-    require_finite("bulk_coeffs_isentropic.v_n_sq", coeffs.v_n_sq)
-    require_finite("bulk_coeffs_isentropic.dμB_dT_sigma", coeffs.dμB_dT_sigma)
-
-    length(coeffs.masses) == 3 || throw(ArgumentError("bulk_coeffs_isentropic.masses must have length 3"))
-    length(coeffs.dM_dT) == 3 || throw(ArgumentError("bulk_coeffs_isentropic.dM_dT must have length 3"))
-    length(coeffs.dM_dμB) == 3 || throw(ArgumentError("bulk_coeffs_isentropic.dM_dμB must have length 3"))
-
-    all(isfinite, coeffs.masses) || throw(ArgumentError("bulk_coeffs_isentropic.masses must be finite"))
-    if !all(m -> m >= 0.0, coeffs.masses)
-        @warn "bulk_coeffs_isentropic.masses contains negative value(s) — may indicate unphysical gap solution; proceeding with abs(m)" masses=coeffs.masses
-    end
-    all(isfinite, coeffs.dM_dT) || throw(ArgumentError("bulk_coeffs_isentropic.dM_dT must be finite"))
-    all(isfinite, coeffs.dM_dμB) || throw(ArgumentError("bulk_coeffs_isentropic.dM_dμB must be finite"))
-    return nothing
-end
-
-@inline function _validate_transport_inputs(
-    quark_params::NamedTuple,
-    thermo_params::NamedTuple,
-    tau::NamedTuple,
-    config::TransportIntegrationConfig;
-    provider=nothing,
-    charges::Union{Nothing,NamedTuple}=nothing,
-    bulk_coeffs_isentropic::Union{Nothing,NamedTuple}=nothing,
-)
-    _validate_quark_thermo_inputs(quark_params, thermo_params; provider=provider)
-    _validate_tau_namedtuple(tau)
-
-    require_positive_integer("integration p_nodes", config.p_nodes)
-    require_positive_integer("integration cos_nodes", config.cos_nodes)
-    require_positive_finite("integration p_max", config.p_max)
-
-    if charges !== nothing
-        require_finite("charges.u", charges.u)
-        require_finite("charges.d", charges.d)
-        require_finite("charges.s", charges.s)
-    end
-
-    if bulk_coeffs_isentropic !== nothing
-        _validate_bulk_coeffs_isentropic(bulk_coeffs_isentropic)
-    end
-
-    return nothing
-end
 
 @inline function _integrate_species_sum(
     term_for_species,
@@ -1153,18 +1081,6 @@ function bulk_viscosity_isentropic(
         provider=provider,
         kwargs...
     )
-end
-
-@inline function _validate_conserved_charge_symbol(charge::Symbol)::Symbol
-    charge in _CONSERVED_CHARGES || error("Unknown conserved charge: $charge. Allowed charges: $(_CONSERVED_CHARGES)")
-    return charge
-end
-
-function _validate_diffusion_background(densities::NamedTuple, pressure::Real, energy::Real)
-    charge_densities = conserved_charge_densities(densities)
-    h = enthalpy_density(pressure, energy)
-    h > 0.0 || error("enthalpy density must be > 0")
-    return charge_densities, h
 end
 
 @inline function _kappa_projection(species::Symbol, charge::Symbol, charge_densities::NamedTuple, enthalpy::Float64, E::Float64)::Float64
